requirements.txt
streamlit
pillow
groq
pandas
python-dotenv
supabase